requirements.txt
# Core framework
fastapi[all]==0.128.0
uvicorn[standard]==0.40.0

# Database
sqlalchemy==2.0.45

# Auth & security
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
bcrypt==4.1.3

# Caching / rate limiting
redis==7.1.0

# HTTP / async
httpx==0.28.1
anyio==4.12.0

# Environment variables
python-dotenv==1.2.1

# Validation
email-validator==2.3.0

# Testing
pytest==9.0.2
pytest-cov==7.0.0
pytest-asyncio==1.3.0

# Utilities
pydantic==2.12.5
pydantic-settings==2.12.0
